pyproject.toml
[project]
name = "gramps-mcp"
version = "1.1.0"
description = "AI-Powered Genealogy Research & Management - MCP server for Gramps Web API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "httpx>=0.25.0",
    "markdownify>=0.11.6",
    "mcp>=1.2.0",
    "pydantic>=2.0.0",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.23.0",
]

[tool.ruff]
exclude = [
    "examples/",
    "venv/",
    ".git/",
    "__pycache__/",
    "*.egg-info/",
    ".venv/",
    "tests/"
]

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "black>=23.0.0",
    "mypy>=1.5.0",
    "pip-audit>=2.6.0",
    "pre-commit>=4.3.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "requests>=2.31.0",
    "ruff>=0.1.0",
    "tdd-guard-pytest>=0.1.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "slow: marks tests as slow running",
    "timeout: marks tests with specific timeout values",
    "server: marks tests requiring a running MCP server (skipped by default)",
]
addopts = "--verbose --strict-markers --tb=short --cov=src --cov-branch --cov-report=term-missing --cov-fail-under=80 -m 'not server'"
asyncio_mode = "auto"
